feat: ensure not call DeleteFile when file does not exists and SaveUserPicture throws

+++ src/domain/usecases/change-profile-picture.ts
@@ -31,7 +31,8 @@ export const setupChangeProfilePicture: Setup =
     try {
       await userProfileRepo.savePicture(userProfile)
     } catch (error) {
-      await fileStorage.delete({ key })
+      if (file !== undefined) await fileStorage.delete({ key })
+      throw error
     }
     return userProfile
   }
